+++ src/main/java/open/commons/core/prog/StrLenRvrOrderingEntry.java
@@ -40,14 +40,14 @@ public class StrLenRvrOrderingEntry implements Comparable<StrLenRvrOrderingEntry
     private final String value;
 
     // 아래 내용에 적용됨.
-    // - key = k.trim();
+    // - key = k.strip();
     // [PATCH] JDK 표준 API의 JSpecify 미지원 '우회용' 어노테이션.
     // [TODO] 향후 JDK 자체 지원 또는 외부 Stub 환경이 갖춰지면 '제거'
     @SuppressWarnings("null")
     public StrLenRvrOrderingEntry(String k, String v) {
         ObjectUtils.requireNonNulls(k, v);
 
-        key = k.trim();
+        key = k.strip();
         len = k.length();
 
         value = v;